[#1599] Consolidate team, org model definitions back to single file
